requirements.txt
ultralytics
streamlit
opencv-python
pillow
numpy
torch
torchvision